fix #157 refactor: リファクタリング

+++ src/app/services/database-books.service.ts
@@ -6,6 +6,7 @@ import { Observable } from 'rxjs';
 import { firestore } from 'firebase/app';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { map } from 'rxjs/operators';
+import { AngularFireFunctions } from '@angular/fire/functions';
 
 @Injectable({
   providedIn: 'root',
@@ -14,10 +15,11 @@ export class DatabaseBooksService {
   constructor(
     private db: AngularFirestore,
     private authService: AuthService,
-    private snackBar: MatSnackBar
+    private snackBar: MatSnackBar,
+    private fns: AngularFireFunctions
   ) {}
 
-  createToFavoriteBook(book: Book): Promise<void> {
+  async createToFavoriteBook(book: Book): Promise<void> {
     const uid = this.authService.uid;
     return this.db
       .doc(`users/${this.authService.uid}/favoriteBooks/${book.id}`)
@@ -34,7 +36,7 @@ export class DatabaseBooksService {
       });
   }
 
-  getToFavoriteBooks(sort, order): Observable<Book[]> {
+  getToFavoriteBooks(sort: string, order: any): Observable<Book[]> {
     return this.db
       .collection<Book>(
         `users/${this.authService.uid}/favoriteBooks`,
@@ -45,7 +47,7 @@ export class DatabaseBooksService {
       .valueChanges();
   }
 
-  getToFavoriteBook(id): Observable<Book> {
+  getToFavoriteBook(id: string): Observable<Book> {
     return this.db
       .doc<Book>(`users/${this.authService.uid}/favoriteBooks/${id}`)
       .valueChanges();
@@ -58,9 +60,13 @@ export class DatabaseBooksService {
       .pipe(map((books) => books.map((book) => book.id)));
   }
 
-  removeToFavoriteBook(id): Promise<void> {
-    return this.db
-      .doc(`users/${this.authService.uid}/favoriteBooks/${id}`)
-      .delete();
+  async removeToFavoriteBook(id: string): Promise<void> {
+    this.db.doc(`users/${this.authService.uid}/favoriteBooks/${id}`).delete();
+    return this.deleteReviews(id);
+  }
+
+  async deleteReviews(id: string): Promise<void> {
+    const callable = this.fns.httpsCallable('deleteReviews');
+    return callable(id).toPromise();
   }
 }
